+++ src/test/java/kitchenpos/acceptance/order/TableAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance.table;
+package kitchenpos.acceptance.order;
 
 import static org.assertj.core.api.Assertions.assertThat;
